feat: 9 and 26 issues
